Update script.js
Continued the guided project while trying to pseudocode out loud between major developments. Planning to have a separate md file to practice pseudocode, but written down. Less hesitation towards Math objects, but somewhat confused on the syntax between else if statements and Logic OR. Vague idea where the "weapons[currentWeapon].name" property(?) came from/significance but hoping for a more detailed exposure to it in the future. Otherwise getting a good handle on Javascript's effect on the DOM, using arrays, recognizing functions, parameters, scope of variables, and so forth.