Update README.md
"moved notes to README to make it easier to check for mistakes as the code wasn't executing as expected compared to in the fCC project window. Also testing if the code will properly update from VS to Github." - Test failed but code is executing how it should be, it just wasn't customized to look anything but plain, aside from the buttons, hence the blinding white background.